Added delete skill

+++ src/main/java/com/netcracker/skillstable/service/dto/SkillService.java
@@ -2,8 +2,7 @@ package com.netcracker.skillstable.service.dto;
 
 import com.netcracker.skillstable.model.EAVObject;
 import com.netcracker.skillstable.model.Parameter;
-import com.netcracker.skillstable.model.dto.Skill;
-import com.netcracker.skillstable.model.dto.Team;
+import com.netcracker.skillstable.model.dto.*;
 import com.netcracker.skillstable.service.EAVService;
 import com.netcracker.skillstable.service.MetamodelService;
 import com.netcracker.skillstable.service.converter.SkillConverter;
@@ -20,6 +19,8 @@ public class SkillService {
     private EAVService eavService;
     @Autowired
     private MetamodelService metamodelService;
+    @Autowired
+    private UserService userService;
 
 
     public Skill createSkill(Skill skill) {
@@ -51,6 +52,16 @@ public class SkillService {
     }
 
     public void deleteSkill(Integer skillId) {
+        Optional<Skill> optionalSkill = this.getSkillById(skillId);
+        if (!optionalSkill.isPresent()) {
+            return;
+        }
+
+        Skill skill = optionalSkill.get();
+        for (User user : userService.getAllUsers()) {
+            user.deleteSkillLevel(skill);
+        }
+        
         eavService.deleteEAVObj(skillId);
     }
 }
